refactor: modularize browser CDP scripts
