+++ package.json
@@ -17,6 +17,7 @@
     "release": "changeset publish",
     "check:changed": "bun run scripts/check-changed.mjs",
     "refactor:compare": "node scripts/refactor-compare.mjs",
+    "browser-cdp:sync": "node scripts/sync-browser-cdp.mjs --write",
     "knip": "knip --workspace @howaboua/pi-codex-conversion --workspace @howaboua/pi-gippity-control --workspace @howaboua/pi-auto-trees --workspace @howaboua/pi-pet --workspace @howaboua/pi-subagent-review --workspace @howaboua/pi-smart-btw --workspace @howaboua/pi-shepherdr --no-progress --include files,dependencies,unlisted,unresolved,cycles,exports,nsExports,types,nsTypes,enumMembers,namespaceMembers,duplicates",
     "changeset:aggregates": "bun run scripts/add-aggregate-changesets.mjs",
     "hooks:install": "git config core.hooksPath .githooks",
