+++ packages/pi-codex-conversion/examples/custom-tools/browser/cdp-lib/README.md
@@ -1,0 +1,3 @@
+# Browser CDP internals
+
+This directory is the canonical generic implementation. After editing it, run `bun browser-cdp:sync` from the repository root to refresh the Pi Codex example copy. The package check rejects drift.
