+++ packages/pi-codex-conversion/examples/custom-tools/browser/cdp-lib/runtime.mjs
@@ -1,0 +1,160 @@
+import { existsSync, mkdirSync, readFileSync } from 'fs';
+import http from 'http';
+import { homedir } from 'os';
+import { resolve } from 'path';
+
+const TIMEOUT = 15000;
+const NAVIGATION_TIMEOUT = 30000;
+const IDLE_TIMEOUT = 20 * 60 * 1000;
+const DAEMON_CONNECT_DELAY = 300;
+const DAEMON_CONNECT_TIMEOUT = TIMEOUT + 2000;
+const MIN_TARGET_PREFIX_LEN = 8;
+const IS_WINDOWS = process.platform === 'win32';
+if (!IS_WINDOWS) process.umask(0o077);
+const RUNTIME_DIR = IS_WINDOWS
+  ? resolve(process.env.LOCALAPPDATA || resolve(homedir(), 'AppData', 'Local'), 'cdp')
+  : process.env.XDG_RUNTIME_DIR
+    ? resolve(process.env.XDG_RUNTIME_DIR, 'cdp')
+    : resolve(homedir(), '.cache', 'cdp');
+try { mkdirSync(RUNTIME_DIR, { recursive: true, mode: 0o700 }); } catch {}
+const PAGES_CACHE = resolve(RUNTIME_DIR, 'pages.json');
+
+function sockPath(targetId) {
+  return IS_WINDOWS
+    ? `\\\\.\\pipe\\cdp-${targetId}`
+    : resolve(RUNTIME_DIR, `cdp-${targetId}.sock`);
+}
+
+async function getJson(url, timeout = 2000) {
+  return new Promise((resolve, reject) => {
+    const req = http.get(url, { timeout }, (res) => {
+      let body = '';
+      res.setEncoding('utf8');
+      res.on('data', chunk => { body += chunk; });
+      res.on('end', () => {
+        if (res.statusCode < 200 || res.statusCode >= 300) {
+          reject(new Error(`HTTP ${res.statusCode} from ${url}`));
+          return;
+        }
+        try { resolve(JSON.parse(body)); } catch (err) { reject(err); }
+      });
+    });
+    req.on('timeout', () => req.destroy(new Error(`Timeout fetching ${url}`)));
+    req.on('error', reject);
+  });
+}
+
+async function probeDebugPort(port, host = process.env.CDP_HOST || '127.0.0.1') {
+  const version = await getJson(`http://${host}:${port}/json/version`);
+  if (!version.webSocketDebuggerUrl) throw new Error(`No webSocketDebuggerUrl from ${host}:${port}`);
+  return version.webSocketDebuggerUrl;
+}
+
+async function getWsUrl(recovery = 'Enable remote debugging or set CDP_PORT/CDP_PORT_FILE.') {
+  const home = homedir();
+  // macOS: ~/Library/Application Support/<name>/DevToolsActivePort
+  const macBrowsers = [
+    'Google/Chrome', 'Google/Chrome Beta', 'Google/Chrome for Testing',
+    'Chromium', 'BraveSoftware/Brave-Browser', 'Microsoft Edge',
+  ];
+  // Linux: ~/.config/<name>/DevToolsActivePort
+  const linuxBrowsers = [
+    'google-chrome', 'google-chrome-beta', 'chromium',
+    'vivaldi', 'vivaldi-snapshot',
+    'BraveSoftware/Brave-Browser', 'microsoft-edge',
+  ];
+  // Linux Flatpak: ~/.var/app/<app-id>/config/<name>/DevToolsActivePort
+  const flatpakBrowsers = [
+    ['org.chromium.Chromium', 'chromium'],
+    ['com.google.Chrome', 'google-chrome'],
+    ['com.brave.Browser', 'BraveSoftware/Brave-Browser'],
+    ['com.microsoft.Edge', 'microsoft-edge'],
+    ['com.vivaldi.Vivaldi', 'vivaldi'],
+  ];
+  const candidates = [
+    process.env.CDP_PORT_FILE,
+    ...macBrowsers.flatMap(b => [
+      resolve(home, 'Library/Application Support', b, 'DevToolsActivePort'),
+      resolve(home, 'Library/Application Support', b, 'Default/DevToolsActivePort'),
+    ]),
+    ...linuxBrowsers.flatMap(b => [
+      resolve(home, '.config', b, 'DevToolsActivePort'),
+      resolve(home, '.config', b, 'Default/DevToolsActivePort'),
+    ]),
+    ...flatpakBrowsers.flatMap(([appId, name]) => [
+      resolve(home, '.var/app', appId, 'config', name, 'DevToolsActivePort'),
+      resolve(home, '.var/app', appId, 'config', name, 'Default/DevToolsActivePort'),
+    ]),
+    // Windows: %LOCALAPPDATA%/<name>/User Data/DevToolsActivePort
+    ...(IS_WINDOWS ? ['Google/Chrome', 'BraveSoftware/Brave-Browser', 'Microsoft/Edge'].flatMap(b => {
+      const base = process.env.LOCALAPPDATA || resolve(home, 'AppData/Local');
+      return [
+        resolve(base, b, 'User Data/DevToolsActivePort'),
+        resolve(base, b, 'User Data/Default/DevToolsActivePort'),
+      ];
+    }) : []),
+  ].filter(Boolean);
+  const host = process.env.CDP_HOST || '127.0.0.1';
+
+  // Prefer a fixed debugging port for deterministic one-shot access.
+  // This gives agents deterministic one-shot access to the logged-in browser.
+  const ports = [process.env.CDP_PORT || '9222'];
+  const errors = [];
+  for (const port of ports) {
+    try { return await probeDebugPort(port, host); }
+    catch (err) { errors.push(`${host}:${port} ${err.message}`); }
+  }
+
+  // Fallback for browsers that expose a dynamic debugging port through
+  // DevToolsActivePort instead of the fixed 9222 endpoint.
+  const portFile = candidates.find(p => existsSync(p));
+  if (portFile) {
+    const lines = readFileSync(portFile, 'utf8').trim().split('\n');
+    if (lines.length < 2 || !lines[0] || !lines[1]) throw new Error(`Invalid DevToolsActivePort file: ${portFile}`);
+    return `ws://${host}:${lines[0]}${lines[1]}`;
+  }
+
+  throw new Error(`CDP HTTP discovery failed and no DevToolsActivePort found. ${recovery} Tried: ${errors.join('; ')}`);
+}
+
+const sleep = (ms) => new Promise(r => setTimeout(r, ms));
+
+function resolvePrefix(prefix, candidates, noun = 'target', missingHint = '') {
+  const upper = prefix.toUpperCase();
+  const matches = candidates.filter(candidate => candidate.toUpperCase().startsWith(upper));
+  if (matches.length === 0) {
+    const hint = missingHint ? ` ${missingHint}` : '';
+    throw new Error(`No ${noun} matching prefix "${prefix}".${hint}`);
+  }
+  if (matches.length > 1) {
+    throw new Error(`Ambiguous prefix "${prefix}" — matches ${matches.length} ${noun}s. Use more characters.`);
+  }
+  return matches[0];
+}
+
+function getDisplayPrefixLength(targetIds) {
+  if (targetIds.length === 0) return MIN_TARGET_PREFIX_LEN;
+  const maxLen = Math.max(...targetIds.map(id => id.length));
+  for (let len = MIN_TARGET_PREFIX_LEN; len <= maxLen; len++) {
+    const prefixes = new Set(targetIds.map(id => id.slice(0, len).toUpperCase()));
+    if (prefixes.size === targetIds.length) return len;
+  }
+  return maxLen;
+}
+
+export {
+  DAEMON_CONNECT_DELAY,
+  DAEMON_CONNECT_TIMEOUT,
+  IDLE_TIMEOUT,
+  IS_WINDOWS,
+  NAVIGATION_TIMEOUT,
+  PAGES_CACHE,
+  RUNTIME_DIR,
+  TIMEOUT,
+  getDisplayPrefixLength,
+  getWsUrl,
+  probeDebugPort,
+  resolvePrefix,
+  sleep,
+  sockPath,
+};
